requirements.txt
Flask
langchain
langchain-openai
openai
python-dotenv
Flask-CORS
Flask-Limiter
